+++ storm-core/src/jvm/org/apache/storm/Config.java
@@ -1539,15 +1539,15 @@ public class Config extends HashMap<String, Object> {
     public static final String TOPOLOGY_COMPONENT_CPU_PCORE_PERCENT = "topology.component.cpu.pcore.percent";
 
     /**
-     * The class name of the {@link backtype.storm.state.StateProvider} implementation. If not specified
-     * defaults to {@link backtype.storm.state.InMemoryKeyValueStateProvider}. This can be overridden
+     * The class name of the {@link org.apache.storm.state.StateProvider} implementation. If not specified
+     * defaults to {@link org.apache.storm.state.InMemoryKeyValueStateProvider}. This can be overridden
      * at the component level.
      */
     @isString
     public static final String TOPOLOGY_STATE_PROVIDER = "topology.state.provider";
 
     /**
-     * The configuration specific to the {@link backtype.storm.state.StateProvider} implementation.
+     * The configuration specific to the {@link org.apache.storm.state.StateProvider} implementation.
      * This can be overridden at the component level. The value and the interpretation of this config
      * is based on the state provider implementation. For e.g. this could be just a config file name
      * which contains the config for the state provider implementation.
@@ -1557,7 +1557,7 @@ public class Config extends HashMap<String, Object> {
 
     /**
      * Topology configuration to specify the checkpoint interval (in millis) at which the
-     * topology state is saved when {@link backtype.storm.topology.IStatefulBolt} bolts are involved.
+     * topology state is saved when {@link org.apache.storm.topology.IStatefulBolt} bolts are involved.
      */
     @isInteger
     @isPositiveNumber
