+++ storm-core/src/jvm/org/apache/storm/state/InMemoryKeyValueState.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package backtype.storm.state;
+package org.apache.storm.state;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
