+++ storm-core/test/jvm/org/apache/storm/topology/StatefulBoltExecutorTest.java
@@ -15,16 +15,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package backtype.storm.topology;
+package org.apache.storm.topology;
 
-import backtype.storm.generated.GlobalStreamId;
-import backtype.storm.generated.Grouping;
-import backtype.storm.spout.CheckpointSpout;
-import backtype.storm.state.KeyValueState;
-import backtype.storm.state.State;
-import backtype.storm.task.OutputCollector;
-import backtype.storm.task.TopologyContext;
-import backtype.storm.tuple.Tuple;
+import org.apache.storm.generated.GlobalStreamId;
+import org.apache.storm.generated.Grouping;
+import org.apache.storm.spout.CheckpointSpout;
+import org.apache.storm.state.KeyValueState;
+import org.apache.storm.state.State;
+import org.apache.storm.task.OutputCollector;
+import org.apache.storm.task.TopologyContext;
+import org.apache.storm.tuple.Tuple;
+import org.apache.storm.topology.IStatefulBolt;
+import org.apache.storm.topology.StatefulBoltExecutor;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
@@ -35,8 +37,8 @@ import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 
-import static backtype.storm.spout.CheckPointState.Action.*;
-import static backtype.storm.spout.CheckpointSpout.*;
+import static org.apache.storm.spout.CheckPointState.Action.*;
+import static org.apache.storm.spout.CheckpointSpout.*;
 
 /**
  * Unit tests for {@link StatefulBoltExecutor}
